separate ExperimentResults
