+++ tabrepo/benchmark/result/__init__.py
@@ -1,3 +1,4 @@
 from .ag_bag_result import AGBagResult
 from .baseline_result import BaselineResult
 from .config_result import ConfigResult
+from .experiment_results import ExperimentResults
